requirements.txt
python-dotenv
groq
gradio
speechrecognition
gtts
pygame





